Changed the Int64 to Double
The Int64 was really difficult to use when it comes to the Slider component. thus I have decided to change the type to double

+++ ReadingNote/DataGroup/DataController.swift
@@ -23,7 +23,7 @@ class DataController: ObservableObject {
         }
     }
     
-    func addNote(name: String, readingTime: Int64, context: NSManagedObjectContext) {
+    func addNote(name: String, readingTime: Double, context: NSManagedObjectContext) {
         let book = Book(context: context)
         book.name = name
         book.readingTime = readingTime
@@ -33,7 +33,7 @@ class DataController: ObservableObject {
         save(context: context)
     }
     
-    func editNote(book: Book, name: String, readingTime: Int64, context: NSManagedObjectContext) {
+    func editNote(book: Book, name: String, readingTime: Double, context: NSManagedObjectContext) {
         book.name = name
         book.readingTime = readingTime
         book.readDate = Date()
